Backend/requirements.txt
# ============================================================================
# SmartEyeSsen Backend - Python Dependencies
# ============================================================================
# FastAPI 프레임워크 및 웹 서버
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================================================
# 데이터베이스 및 ORM
# ============================================================================
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0  # pymysql 암호화 지원
alembic==1.13.1  # 데이터베이스 마이그레이션 도구

# ============================================================================
# 데이터 검증 및 설정
# ============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================================================
# AI/ML 라이브러리
# ============================================================================
# PyTorch (DocLayout-YOLO 호환 버전)
torch==2.0.1
torchvision==0.15.2

# DocLayout-YOLO (Project 검증된 버전)
ultralytics==8.0.196

# PaddleOCR
paddlepaddle==2.6.0
paddleocr==2.7.0

# Tesseract OCR
pytesseract==0.3.10

# Hugging Face Transformers
huggingface-hub>=0.17.0
transformers==4.35.2

# OpenAI API (AI 설명 생성)
openai==1.10.0

# ============================================================================
# 이미지 처리
# ============================================================================
pillow==10.2.0
opencv-python==4.9.0.80
matplotlib>=3.5.0
scikit-image==0.22.0
imageio==2.31.6

# ============================================================================
# 데이터 처리 및 분석
# ============================================================================
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
numpy==1.26.4

# ============================================================================
# 문서 생성 및 처리
# ============================================================================
python-docx==1.1.0
PyMuPDF==1.23.8  # PDF 처리

# ============================================================================
# 유틸리티
# ============================================================================
# 날짜/시간 처리
python-dateutil==2.8.2
# HTTP 클라이언트
httpx==0.26.0
# 로깅
loguru==0.7.2

# ============================================================================
# 보안 (선택사항)
# ============================================================================
# JWT 토큰 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# 개발 도구 (선택사항)
# ============================================================================
# pytest==7.4.4
# pytest-asyncio==0.23.3
# black==23.12.1
# flake8==7.0.0
